--- modules/task_2/stepanov_hoar_sort_with_odd_even_batcher_merge_omp/hoar_sort_with_odd_even_batcher_merge.h
@@ -1,11 +1,0 @@
-// Copyright 2022 George Amberson
-#ifndef MODULES_TASK_2_STEPANOV_HOAR_SORT_WITH_ODD_EVEN_BATCHER_MERGE_OMP_HOAR_SORT_WITH_ODD_EVEN_BATCHER_MERGE_H_
-#define MODULES_TASK_2_STEPANOV_HOAR_SORT_WITH_ODD_EVEN_BATCHER_MERGE_OMP_HOAR_SORT_WITH_ODD_EVEN_BATCHER_MERGE_H_
-#include <omp.h>
-#include <vector>
-#include <list>
-#include <algorithm>
-std::vector<float>* create_random_vector(int size_n);
-void hoar_sort_with_batch_merge(std::vector<float>* vec);
-void hoar_sort_with_batch_merge_omp(std::vector<float>* vec);
-#endif  // MODULES_TASK_2_STEPANOV_HOAR_SORT_WITH_ODD_EVEN_BATCHER_MERGE_OMP_HOAR_SORT_WITH_ODD_EVEN_BATCHER_MERGE_H_
